requirements.txt
ultralytics
python-chess